+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # 0.12.0
 - Added PropertyListAddMode for easy message compose
+- Added Option MatchUnsafe
+- Change: GetOrElse remaned to GetOrDefault and can return null value
 
 # 0.11.0
 - Initial Memoize implementation